Ch. 9 - auth system to sep. module.
Full app is working, subject to a couple of changes from book code as noted in comments.

  Q. Why were these hacks/changes necessary?
  A. At a guess, maybe restas src changed since the book was written (a guess).

N.B. To run the app from this commit now requires presence of auth system as a separate
component, now found in the new repo 'authdemo'. That should be present in quicklisp/local-projects.

This, ch.9., is to demonstrate & explore refactoring to components in restas.

(To run the app without need for that separate module, just checkout the commit before this one).